api cleanup, removed parser feature
